requirements.txt
chromadb==1.5.5
langchain_chroma==1.1.0
langchain_community==0.4.1
langchain_core==1.2.23
langchain_huggingface==1.2.1
langchain_text_splitters==1.1.1
python-dotenv==1.2.2
sentence_transformers==5.3.0
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
